Add a loading bar service for state changes
Created a reusable "stateLoadingBar" provider to show an application loading bar on state transitions. This is a small wrapper around the "angular-loading-bar" 3rd party component. This wrapper was necessary, in order to add functionality to prevent the loading bar from showing on very fast route changes (route changes without resolves, for instance). The wrapper allows you to set a minimum "loadingThreshold", in order to prevent the display of the loading bar in these scenarios.